fix: init nft table structure in route manager
Flush and delete nft table and release netlink connection
when nft.initStructure() failed.

Related: #276
